fix(desktop): stabilize flaky bun tests
