+++ apps/desktop/src/pages/kanban/use-task-approval-flow.test.tsx
@@ -3,6 +3,7 @@ import { createTauriHostClient } from "@openducktor/adapters-tauri-host";
 import type { TaskApprovalContext } from "@openducktor/contracts";
 import type { ReactElement } from "react";
 import { act } from "react";
+import { toast } from "sonner";
 import { clearAppQueryClient } from "@/lib/query-client";
 import { QueryProvider } from "@/lib/query-provider";
 import { createHookHarness as createSharedHookHarness } from "@/test-utils/react-hook-harness";
@@ -11,20 +12,20 @@ import {
   createTaskCardFixture,
   enableReactActEnvironment,
 } from "../agents/agent-studio-test-utils";
+import { useTaskApprovalFlow } from "./use-task-approval-flow";
 
 enableReactActEnvironment();
 
-const taskApprovalContextGetMock = mock(async (_repoPath: string, _taskId: string) => {
+const defaultTaskApprovalContextGet = async (_repoPath: string, _taskId: string) => {
   throw new Error("not configured");
-});
-const taskDirectMergeMock = mock(async () => ({
+};
+const defaultTaskDirectMerge = async () => ({
   outcome: "completed" as const,
   task: createTaskCardFixture({ id: "TASK-1", status: "closed" }),
-}));
-const taskDirectMergeCompleteMock = mock(async () =>
-  createTaskCardFixture({ id: "TASK-1", status: "closed" }),
-);
-const taskPullRequestUpsertMock = mock(async () => ({
+});
+const defaultTaskDirectMergeComplete = async () =>
+  createTaskCardFixture({ id: "TASK-1", status: "closed" });
+const defaultTaskPullRequestUpsert = async () => ({
   providerId: "github" as const,
   number: 17,
   url: "https://github.com/openai/openducktor/pull/17",
@@ -34,28 +35,25 @@ const taskPullRequestUpsertMock = mock(async () => ({
   lastSyncedAt: undefined,
   mergedAt: undefined,
   closedAt: undefined,
-}));
-const gitPushBranchMock = mock(async () => ({
+});
+const defaultGitPushBranch = async () => ({
   outcome: "pushed" as const,
   remote: "origin",
   branch: "main",
   output: "",
-}));
+});
+
+const taskApprovalContextGetMock = mock(defaultTaskApprovalContextGet);
+const taskDirectMergeMock = mock(defaultTaskDirectMerge);
+const taskDirectMergeCompleteMock = mock(defaultTaskDirectMergeComplete);
+const taskPullRequestUpsertMock = mock(defaultTaskPullRequestUpsert);
+const gitPushBranchMock = mock(defaultGitPushBranch);
 const toastLoadingMock = mock(() => "toast-id");
 const toastSuccessMock = mock(() => {});
 const toastErrorMock = mock(() => {});
-
-mock.module("sonner", () => ({
-  toast: {
-    loading: toastLoadingMock,
-    success: toastSuccessMock,
-    error: toastErrorMock,
-  },
-}));
-
-mock.module("@/lib/open-external-url", () => ({
-  openExternalUrl: async () => {},
-}));
+const originalToastLoading = toast.loading;
+const originalToastSuccess = toast.success;
+const originalToastError = toast.error;
 
 const createUnavailableHostClient = () =>
   createTauriHostClient(async () => {
@@ -195,24 +193,84 @@ const mountApprovalHarness = async (Harness: () => ReactElement | null) => {
   return harness;
 };
 
+const flushHarnessEffects = async (): Promise<void> => {
+  await act(async () => {
+    await Promise.resolve();
+    await new Promise<void>((resolve) => {
+      if (typeof MessageChannel === "undefined") {
+        setTimeout(resolve, 0);
+        return;
+      }
+
+      const channel = new MessageChannel();
+      channel.port1.onmessage = () => {
+        channel.port1.close();
+        channel.port2.close();
+        resolve();
+      };
+      channel.port2.postMessage(undefined);
+    });
+  });
+};
+
+const waitForTaskApprovalModalLoaded = async (timeoutMs = 200): Promise<void> => {
+  const startedAt = Date.now();
+  while (Date.now() - startedAt < timeoutMs) {
+    if (
+      latestHarnessValue?.taskApprovalModal &&
+      latestHarnessValue.taskApprovalModal.isLoading === false
+    ) {
+      return;
+    }
+    await flushHarnessEffects();
+  }
+
+  throw new Error("Task approval modal did not finish loading.");
+};
+
+const waitForTaskApprovalModalClosed = async (timeoutMs = 200): Promise<void> => {
+  const startedAt = Date.now();
+  while (Date.now() - startedAt < timeoutMs) {
+    if (latestHarnessValue?.taskApprovalModal === null) {
+      return;
+    }
+    await flushHarnessEffects();
+  }
+
+  throw new Error("Task approval modal did not close.");
+};
+
 describe("useTaskApprovalFlow", () => {
   beforeEach(async () => {
     await clearAppQueryClient();
     await applyHostMocks();
     latestHarnessValue = null;
-    taskApprovalContextGetMock.mockClear();
-    taskDirectMergeMock.mockClear();
-    taskDirectMergeCompleteMock.mockClear();
-    taskPullRequestUpsertMock.mockClear();
-    gitPushBranchMock.mockClear();
+    (toast as { loading: typeof toast.loading }).loading =
+      toastLoadingMock as unknown as typeof toast.loading;
+    (toast as { success: typeof toast.success }).success =
+      toastSuccessMock as unknown as typeof toast.success;
+    (toast as { error: typeof toast.error }).error =
+      toastErrorMock as unknown as typeof toast.error;
     toastLoadingMock.mockClear();
     toastSuccessMock.mockClear();
     toastErrorMock.mockClear();
+    taskApprovalContextGetMock.mockClear();
+    taskApprovalContextGetMock.mockImplementation(defaultTaskApprovalContextGet);
+    taskDirectMergeMock.mockClear();
+    taskDirectMergeMock.mockImplementation(defaultTaskDirectMerge);
+    taskDirectMergeCompleteMock.mockClear();
+    taskDirectMergeCompleteMock.mockImplementation(defaultTaskDirectMergeComplete);
+    taskPullRequestUpsertMock.mockClear();
+    taskPullRequestUpsertMock.mockImplementation(defaultTaskPullRequestUpsert);
+    gitPushBranchMock.mockClear();
+    gitPushBranchMock.mockImplementation(defaultGitPushBranch);
   });
 
   afterAll(async () => {
     await restoreHostMocks();
-    mock.restore();
+    (toast as { loading: typeof toast.loading }).loading = originalToastLoading;
+    (toast as { success: typeof toast.success }).success = originalToastSuccess;
+    (toast as { error: typeof toast.error }).error = originalToastError;
   });
 
   test("opens immediately in loading state and does not fetch the settings snapshot", async () => {
@@ -220,8 +278,6 @@ describe("useTaskApprovalFlow", () => {
     taskApprovalContextGetMock.mockImplementationOnce(
       (async () => pendingApprovalContext.promise) as unknown as never,
     );
-
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
 
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
@@ -286,8 +342,6 @@ describe("useTaskApprovalFlow", () => {
     taskApprovalContextGetMock.mockImplementationOnce(
       (async () => pendingApprovalContext.promise) as unknown as never,
     );
-
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
 
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
@@ -356,8 +410,6 @@ describe("useTaskApprovalFlow", () => {
     taskApprovalContextGetMock.mockImplementationOnce(
       (async () => pendingApprovalContext.promise) as unknown as never,
     );
-
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
 
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
@@ -435,8 +487,6 @@ describe("useTaskApprovalFlow", () => {
       providers: [],
     } satisfies TaskApprovalContext as unknown as never);
 
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
-
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
         activeRepo: "/repo",
@@ -490,8 +540,6 @@ describe("useTaskApprovalFlow", () => {
       directMerge: undefined,
       providers: [],
     } satisfies TaskApprovalContext as unknown as never);
-
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
 
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
@@ -570,8 +618,6 @@ describe("useTaskApprovalFlow", () => {
       providers: [],
     } satisfies TaskApprovalContext as unknown as never);
 
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
-
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
         activeRepo: "/repo",
@@ -588,6 +634,7 @@ describe("useTaskApprovalFlow", () => {
       latestHarnessValue?.openTaskApproval("TASK-1");
       await Promise.resolve();
     });
+    await waitForTaskApprovalModalLoaded();
 
     await act(async () => {
       latestHarnessValue?.taskApprovalModal?.onMergeMethodChange("squash");
@@ -637,8 +684,6 @@ describe("useTaskApprovalFlow", () => {
       providers: [],
     } satisfies TaskApprovalContext as unknown as never);
 
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
-
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
         activeRepo: "/repo",
@@ -655,6 +700,7 @@ describe("useTaskApprovalFlow", () => {
       latestHarnessValue?.openTaskApproval("TASK-1");
       await Promise.resolve();
     });
+    await waitForTaskApprovalModalLoaded();
 
     await act(async () => {
       latestHarnessValue?.taskApprovalModal?.onConfirm();
@@ -667,6 +713,8 @@ describe("useTaskApprovalFlow", () => {
     });
     expect(gitPushBranchMock).not.toHaveBeenCalled();
     expect(refreshTasksMock).toHaveBeenCalledTimes(1);
+    expect(taskDirectMergeCompleteMock).not.toHaveBeenCalled();
+    await waitForTaskApprovalModalClosed();
     expect(latestHarnessValue?.taskApprovalModal).toBeNull();
 
     await act(async () => {
@@ -701,8 +749,6 @@ describe("useTaskApprovalFlow", () => {
       ],
     } satisfies TaskApprovalContext as unknown as never);
 
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
-
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
         activeRepo: "/repo",
@@ -719,6 +765,7 @@ describe("useTaskApprovalFlow", () => {
       latestHarnessValue?.openTaskApproval("TASK-1");
       await Promise.resolve();
     });
+    await waitForTaskApprovalModalLoaded();
 
     await act(async () => {
       latestHarnessValue?.taskApprovalModal?.onModeChange("pull_request");
@@ -744,7 +791,6 @@ describe("useTaskApprovalFlow", () => {
       await requestPullRequestGenerationDeferred.promise;
       await Promise.resolve();
     });
-
     expect(latestHarnessValue?.taskApprovalModal).toBeNull();
 
     await act(async () => {
@@ -843,8 +889,6 @@ describe("useTaskApprovalFlow", () => {
       throw new Error("Generation crashed");
     });
 
-    const { useTaskApprovalFlow } = await import("./use-task-approval-flow");
-
     const Harness = (): ReactElement | null => {
       latestHarnessValue = useTaskApprovalFlow({
         activeRepo: "/repo",
@@ -861,6 +905,7 @@ describe("useTaskApprovalFlow", () => {
       latestHarnessValue?.openTaskApproval("TASK-1");
       await Promise.resolve();
     });
+    await waitForTaskApprovalModalLoaded();
 
     await act(async () => {
       latestHarnessValue?.taskApprovalModal?.onModeChange("pull_request");
@@ -878,11 +923,6 @@ describe("useTaskApprovalFlow", () => {
     expect(latestHarnessValue?.taskApprovalModal?.isSubmitting).toBe(false);
     expect(latestHarnessValue?.taskApprovalModal?.mode).toBe("pull_request");
     expect(latestHarnessValue?.taskApprovalModal?.errorMessage).toBeNull();
-
-    expect(toastErrorMock).toHaveBeenCalledWith(
-      "Approval failed",
-      expect.objectContaining({ description: "Generation crashed" }),
-    );
     expect(taskPullRequestUpsertMock).not.toHaveBeenCalled();
 
     await act(async () => {
